refactor; moved data classes out of loaders
