+++ cognite_toolkit/_cdf_tk/client/data_classes/extendable_cognite_file.py
@@ -1,213 +1,14 @@
-"""These are helper data classes for the load module.
-
-There are three types of classes in this module made with different motivations:
-
-* RawDatabaseTable and RawTableList are CogniteResources following the pattern
-  from the cognite-sdk for RAW databases and tables. They are missing from the cognite-sdk,
-  and are needed to load YAML.
-* LoadableNodes and LoadableEdges are extensions of NodeApplyList and EdgeApplyList
-  to also contain the arguments of the 'client.data_modeling.instances.apply()' method. This enables
-  the user to specify these arguments in the YAML file.
-* DeployResult and DeployResults are storing the output of the .deploy_resources and .clean_resources,
-  which then shows a summary of the operation in the terminal. The DeployResults class is a UserList
-  of DeployResult objects, and is used to print the summary in a table.
-"""
-
 from __future__ import annotations
 
 import itertools
-from abc import ABC
-from dataclasses import dataclass
 from datetime import datetime
-from functools import total_ordering
 from typing import Any, Literal
 
 from cognite.client import CogniteClient
-from cognite.client.data_classes._base import (
-    CogniteResourceList,
-    WriteableCogniteResource,
-    WriteableCogniteResourceList,
-)
-from cognite.client.data_classes.data_modeling import DataModelId, DirectRelationReference, ViewId
+from cognite.client.data_classes._base import CogniteResourceList, WriteableCogniteResourceList
+from cognite.client.data_classes.data_modeling import DirectRelationReference, ViewId
 from cognite.client.data_classes.data_modeling.cdm.v1 import CogniteFile, CogniteFileApply
-from cognite.client.data_classes.data_modeling.core import DataModelingSchemaResource
 from cognite.client.utils._text import to_camel_case
-
-
-@dataclass(frozen=True)
-class FunctionScheduleID:
-    function_external_id: str
-    name: str
-
-    def dump(self, camel_case: bool = True) -> dict[str, Any]:
-        return {
-            "functionExternalId" if camel_case else "function_external_id": self.function_external_id,
-            "name": self.name,
-        }
-
-
-@total_ordering
-@dataclass(frozen=True)
-class RawDatabaseTable(WriteableCogniteResource):
-    db_name: str
-    table_name: str | None = None
-
-    @classmethod
-    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> RawDatabaseTable:
-        return cls(db_name=resource["dbName"], table_name=resource.get("tableName"))
-
-    def dump(self, camel_case: bool = True) -> dict[str, Any]:
-        dumped = {
-            "dbName" if camel_case else "db_name": self.db_name,
-        }
-        if self.table_name is not None:
-            dumped["tableName" if camel_case else "table_name"] = self.table_name
-        return dumped
-
-    def as_write(self) -> RawDatabaseTable:
-        return self
-
-    def __lt__(self, other: object) -> bool:
-        if isinstance(other, RawDatabaseTable):
-            if self.db_name == other.db_name:
-                return (self.table_name or "") < (other.table_name or "")
-            else:
-                return self.db_name < other.db_name
-        else:
-            return NotImplemented
-
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, RawDatabaseTable):
-            return self.db_name == other.db_name and self.table_name == other.table_name
-        else:
-            return NotImplemented
-
-    def __str__(self) -> str:
-        if self.table_name is None:
-            return self.db_name
-        else:
-            return super().__str__()
-
-    def __repr__(self) -> str:
-        if self.table_name is None:
-            return f"{type(self).__name__}(db_name='{self.db_name}')"
-        else:
-            return f"{type(self).__name__}(db_name='{self.db_name}', table_name='{self.table_name}')"
-
-
-class RawTableList(WriteableCogniteResourceList[RawDatabaseTable, RawDatabaseTable]):
-    _RESOURCE = RawDatabaseTable
-
-    def as_write(self) -> CogniteResourceList[RawDatabaseTable]:
-        return self
-
-    def as_db_names(self) -> list[str]:
-        return [table.db_name for table in self.data]
-
-
-class _GraphQLDataModelCore(DataModelingSchemaResource["GraphQLDataModelWrite"], ABC):
-    def __init__(
-        self, space: str, external_id: str, version: str, name: str | None = None, description: str | None = None
-    ) -> None:
-        super().__init__(space=space, external_id=external_id, name=name, description=description)
-        self.version = version
-
-    def as_id(self) -> DataModelId:
-        return DataModelId(space=self.space, external_id=self.external_id, version=self.version)
-
-
-class GraphQLDataModelWrite(_GraphQLDataModelCore):
-    def __init__(
-        self,
-        space: str,
-        external_id: str,
-        version: str,
-        name: str | None = None,
-        description: str | None = None,
-        previous_version: str | None = None,
-        dml: str | None = None,
-    ) -> None:
-        super().__init__(space=space, external_id=external_id, version=version, name=name, description=description)
-        self.previous_version = previous_version
-        self.dml = dml
-
-    @classmethod
-    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> GraphQLDataModelWrite:
-        return cls(
-            space=resource["space"],
-            external_id=resource["externalId"],
-            version=str(resource["version"]),
-            name=resource.get("name"),
-            description=resource.get("description"),
-            previous_version=resource.get("previousVersion"),
-            dml=resource.get("dml"),
-        )
-
-    def as_write(self) -> GraphQLDataModelWrite:
-        return self
-
-
-class GraphQLDataModel(_GraphQLDataModelCore):
-    def __init__(
-        self,
-        space: str,
-        external_id: str,
-        version: str,
-        is_global: bool,
-        last_updated_time: int,
-        created_time: int,
-        description: str | None,
-        name: str | None,
-        views: list[ViewId] | None,
-    ) -> None:
-        super().__init__(space=space, external_id=external_id, version=version, name=name, description=description)
-        self.is_global = is_global
-        self.last_updated_time = last_updated_time
-        self.created_time = created_time
-        self.views = views
-
-    def as_write(self) -> GraphQLDataModelWrite:
-        return GraphQLDataModelWrite(
-            space=self.space,
-            external_id=self.external_id,
-            version=self.version,
-            name=self.name,
-            description=self.description,
-        )
-
-    @classmethod
-    def _load(cls, resource: dict[str, Any], cognite_client: CogniteClient | None = None) -> GraphQLDataModel:
-        return cls(
-            space=resource["space"],
-            external_id=resource["externalId"],
-            version=str(resource["version"]),
-            is_global=resource["isGlobal"],
-            last_updated_time=resource["lastUpdatedTime"],
-            created_time=resource["createdTime"],
-            description=resource.get("description"),
-            name=resource.get("name"),
-            views=[
-                ViewId(space=view["space"], external_id=view["externalId"], version=view.get("version"))
-                for view in resource.get("views", [])
-            ],
-        )
-
-
-class GraphQLDataModelWriteList(CogniteResourceList[GraphQLDataModelWrite]):
-    _RESOURCE = GraphQLDataModelWrite
-
-    def as_ids(self) -> list[DataModelId]:
-        return [model.as_id() for model in self.data]
-
-
-class GraphQLDataModelList(WriteableCogniteResourceList[GraphQLDataModelWrite, GraphQLDataModel]):
-    _RESOURCE = GraphQLDataModel
-
-    def as_write(self) -> GraphQLDataModelWriteList:
-        return GraphQLDataModelWriteList([model.as_write() for model in self.data])
-
-    def as_ids(self) -> list[DataModelId]:
-        return [model.as_id() for model in self.data]
 
 
 class ExtendableCogniteFileApply(CogniteFileApply):
