+++ tests/test_unit/test_cdf_tk/test_loaders/test_group_loader.py
@@ -4,6 +4,7 @@ import pytest
 from _pytest.monkeypatch import MonkeyPatch
 from cognite.client.data_classes import Group, GroupWrite, GroupWriteList
 
+from cognite_toolkit._cdf_tk.client.data_classes.raw import RawDatabaseTable
 from cognite_toolkit._cdf_tk.commands import DeployCommand
 from cognite_toolkit._cdf_tk.loaders import (
     DataSetsLoader,
@@ -16,7 +17,6 @@ from cognite_toolkit._cdf_tk.loaders import (
     ResourceLoader,
     SpaceLoader,
 )
-from cognite_toolkit._cdf_tk.loaders.data_classes import RawDatabaseTable
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig
 from tests.data import LOAD_DATA
 from tests.test_unit.approval_client import ApprovalToolkitClient
